camera_feedback : move class definitions into hpp
